add compoment for bus
新增结伴班车: 订单详情，加入班车地址填写，地址列表，结伴班车无匹配
