+++ passenger/pages/bus/address/index.json
@@ -1,5 +1,5 @@
 {
   "navigationBarBackgroundColor": "#F1604F",
   "navigationBarTextStyle": "white",
-  "navigationBarTitleText": "结伴班车"
+  "navigationBarTitleText": "常用地址"
 }
