--- passenger/pages/bus/address_edit/index.wxss
@@ -1,1 +1,0 @@
-/* pages/bus/index/index.wxss */
