--- passenger/pages/bus/address_list/index.wxml
@@ -1,34 +1,0 @@
-<!--pages/bus/index/index.wxml 结伴班车-->
-<!--pages/my/index/index.wxml-->
-<view class="displayf jscc page-bg">
-    <view class="w710 mt10">
-        <view class="bgw br4 my-box-shadow mb10">
-            <view class="fz14 color-666 border-bottom-ccc h50 boxb">
-                <view class="patrbl15">
-                    <image class="bg-F1604F icon-size"></image>
-                    <view class="fr">
-                        <input placeholder='您从哪儿出发'></input>
-                    </view>
-                </view>
-            </view>
-            <view class="fz14 color-666 h50 boxb">
-                <view class="patrbl15">
-                    <image class="bg-F1604F icon-size" src=""></image>
-                    <view class="fr navigator-choice">
-                        <input placeholder='您想要去哪儿'></input>
-                    </view>
-                </view>
-            </view>
-            <view class="fz14 color-666 h50 boxb">
-                <view class="patrbl15">
-                    <image class="bg-F1604F icon-size" src=""></image>
-                    <view class="fr navigator-choice">
-                        <input placeholder='出发时间'></input>
-                    </view>
-                </view>
-            </view>
-            <button type="default">下一步</button>
-        </view>
-    </view>
-</view>
-
