--- passenger/pages/bus/address_list/index.wxss
@@ -1,1 +1,0 @@
-/* pages/bus/index/index.wxss */
