+++ passenger/pages/bus/bus_list/index.js
@@ -1,4 +1,6 @@
 // pages/bus/index/index.js
+const app = getApp(),
+  util = require('../../../utils/util.js')
 Page({
 
   /**
@@ -7,7 +9,10 @@ Page({
   data: {
   
   },
-
+  go(e) {
+    let url = e.currentTarget.dataset.url
+    util.go(url)
+  },
   /**
    * 生命周期函数--监听页面加载
    */
