+++ passenger/pages/bus/detail/index.wxss
@@ -1,0 +1,130 @@
+/* pages/bus/index/index.wxss */
+/* 
+  latest edit by lewyon
+  code start
+*/
+.bybus-content{
+  padding: 12px 12px 0 12px;
+}
+.bybus-banner{
+  width: 100%;
+  height: 160px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.bybus-banner image{
+  display: block;
+  width: 100%;
+  height: 100%;
+}
+.bybus-search{
+  margin-top: 12px;
+  padding:0 10px 10px 10px; 
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.bybus-search .search-text{
+  flex: 1;
+}
+.bybus-search .search-input{
+  flex: 4;
+  border: 1px solid #ccc;
+  border-radius: 15px;
+  text-indent: 10px;
+  padding-left: 10px;
+  height: 30px;
+  line-height: 30px;
+  font-size: 13px;
+  background: #ccc;
+  color: #fff;
+}
+.bybus-msg{
+  flex-wrap: nowrap;
+  border-bottom:1px solid #CCCCCC;
+  padding:12px 0 12px 0;
+}
+.bg-F1604F{
+  width: 8px;
+  height: 8px;
+  border-radius: 50%;
+  margin-left: 12px;
+}
+.card-avatar{
+  width: 40px;
+  height: 40px;
+}
+.msg-text{
+  flex: 5;
+  flex-wrap: wrap;
+  padding-left: 10px;
+}
+.bybus-content{
+  padding-bottom: 12px;
+}
+.bybus-name{
+  width: 100%;
+  color: #333333;
+  font-size: 15px;
+  font-family: "微软雅黑";
+}
+.bybus-name-type{
+  display: inline-block;
+  background: #F1604F;
+  color: #fff;
+  font-size: 13px;
+  margin-left: 6px;
+  padding: 2px 6px;
+  border-radius: 2px;
+}
+.bybus-tip{
+  font-size: 12px;
+  width: 100%;
+  text-align: left;
+  color: #999999;
+  padding: 3px 0;
+}
+.buybus-join{
+  border-top: 1px solid #CCCCCC;
+  padding: 8px 0;
+}
+.buybus-join-btn{
+  width: 80px;
+  height: 30px;
+  text-align: center;
+  font-size: 14px;
+  color: #666;
+  float: right;
+  margin-right: 12px;
+  border: 1px solid #666;
+}
+.sort-btn{
+  text-align: right;
+  padding-bottom: 12px;
+  margin-top: 12px;
+}
+.sort-btn text{
+  display: inline-block;
+  padding: 4px 14px;
+  border-radius: 8px;
+  background: #000000;
+  color: #fff;
+  margin-right: 12px;
+}
+.searcg-btn{
+  position: absolute;
+  bottom: 12px;
+  left: 0;
+  width: 100%;
+  text-align: center;
+  padding-left: 12px;
+  padding-right: 12px;
+}
+.searcg-btn button{
+  width: 100%;
+  height:53px; 
+  background:rgba(241,96,79,1);
+  border-radius: 2px;
+  color: #fff;
+}
