+++ passenger/pages/bus/join/index.json
@@ -1,5 +1,5 @@
 {
   "navigationBarBackgroundColor": "#F1604F",
   "navigationBarTextStyle": "white",
-  "navigationBarTitleText": "结伴班车"
+  "navigationBarTitleText": "加入班车"
 }
