+++ passenger/pages/bus/mismatch/mismatch.wxss
@@ -1,1 +1,33 @@
 /* pages/bus/mismatch/mismatch.wxss */
+/*edit by lewyon   */
+.main-wrap{
+  height: 100%;
+  display: flex;
+  align-content: center;
+  justify-content:center;
+  flex-wrap: wrap;
+  text-align: center;
+}
+.mismatch-title{
+  width: 100%;
+  color: #333333;
+  font-size: 14px;
+}
+.mismatch-text{
+  width: 100%;
+  color: #999999;
+  font-size: 13px;
+  padding: 4px 0;
+}
+.creat-btn{
+  margin-top: 80px;
+  color: #F1604F;
+  width: 220px;
+  border: 1px solid #F1604F;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: #fff;
+  font-size: 14px;
+  height: 35px;
+}
