--- passenger/pages/bus/order_detail/index.json
@@ -1,5 +1,0 @@
-{
-  "navigationBarBackgroundColor": "#F1604F",
-  "navigationBarTextStyle": "white",
-  "navigationBarTitleText": "结伴班车"
-}
