--- passenger/pages/bus/order_detail/index.wxss
@@ -1,1 +1,0 @@
-/* pages/bus/index/index.wxss */
